# src/SqlAccountRestAPI/assets/yamls/openApiExample.yaml
openapi: 3.1.0
info:
  title: SQL Account Rest API
  version: '1.0'
  description: >
    This is a comprehensive REST API for managing SQL accounts. 
    It provides endpoints for user authentication, customer data management, 
    sales transactions, and stock control. Use this API to integrate your applications 
    with our SQL accounting system.
servers:
  - url: https://sqlaccountingapi.hexasync.dev
...
paths:
...
  /api/stock-items/code/{code}:
    get:
      tags:
        - StockItem
      summary: Retrieves stock items by code
      description: This API fetches stock items that match the provided code.
      parameters:
        - name: code
          in: path
          description: The code to search.
          required: true
          schema:
            type: string
            default: ""
            example: SEMI BOM
        - name: limit
          in: query
          description: "Maximum number of results to return (default: 100)."
          style: form
          schema:
            type: integer
            format: int32
            default: 100
            example: 1
        - name: offset
          in: query
          description: "Offset for pagination (default: 0)."
          style: form
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
      responses:
        '200':
          description: Returns the details of the business object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItem'
              examples:
                default_example:
                  $ref: '#/components/examples/StockItemInfoResponseExample'
        '400':
          description: General error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stock-items/days-ago/{days}:
    get:
      tags:
        - StockItem
      summary: Retrieves stock items from a certain number of days ago
      description: Fetches stock items based on the number of days before the current date.
      parameters:
        - name: days
          in: path
          description: The number of days ago to search.
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            default: 0
            example: 1000
        - name: limit
          in: query
          description: "Maximum number of results to return (default: 100)."
          style: form
          schema:
            type: integer
            format: int32
            default: 100
            example: 1
        - name: offset
          in: query
          description: "Offset for pagination (default: 0)."
          style: form
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
      responses:
        '200':
          description: Returns the details of the business object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItem'
              examples:
                default_example:
                  $ref: '#/components/examples/StockItemInfoResponseExample'
        '400':
          description: General error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stock-items/from-date/{date}:
    get:
      tags:
        - StockItem
      summary: Retrieves stock items from a specific date
      description: Fetches stock items recorded from a given date onward.
      parameters:
        - name: date
          in: path
          description: The date in YYYY-MM-DD format.
          required: true
          style: simple
          schema:
            type: string
            default: ""
            example: 2020-01-01
        - name: limit
          in: query
          description: "Maximum number of results to return (default: 100)."
          style: form
          schema:
            type: integer
            format: int32
            default: 100
            example: 1
        - name: offset
          in: query
          description: "Offset for pagination (default: 0)."
          style: form
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
      responses:
        '200':
          description: Returns the details of the business object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItem'
              examples:
                default_example:
                  $ref: '#/components/examples/StockItemInfoResponseExample'
        '400':
          description: General error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stock-item-templates/code/{code}:
    get:
      tags:
        - StockItemTemplate
      summary: Retrieves stock item templates by code
      description: This API fetches stock item templates that match the provided code.
      parameters:
        - name: code
          in: path
          description: The code to search.
          required: true
          schema:
            type: string
            default: ""
        - name: limit
          in: query
          description: "Maximum number of results to return (example: 100)."
          style: form
          schema:
            type: integer
            format: int32
            default: 100
            example: 1
        - name: offset
          in: query
          description: "Offset for pagination (default: 0)."
          style: form
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
      responses:
        '200':
          description: Returns the details of the business object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItemTemplate'
              examples:
                default_example:
                  $ref: '#/components/examples/StockItemTemplateInfoResponseExample'
        '400':
          description: General error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stock-item-templates/days-ago/{days}:
    get:
      tags:
        - StockItemTemplate
      summary: Retrieves stock item templates from a certain number of days ago
      description: Fetches stock item templates based on the number of days before the current date.
      parameters:
        - name: days
          in: path
          description: The number of days ago to search.
          required: true
          style: simple
          schema:
            type: integer
            format: int32
            default: 0
            example: 1000
        - name: limit
          in: query
          description: "Maximum number of results to return (default: 100)."
          style: form
          schema:
            type: integer
            format: int32
            default: 100
            example: 1
        - name: offset
          in: query
          description: "Offset for pagination (default: 0)."
          style: form
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
      responses:
        '200':
          description: Returns the details of the business object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItemTemplate'
              examples:
                default_example:
                  $ref: '#/components/examples/StockItemTemplateInfoResponseExample'
        '400':
          description: General error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/stock-item-templates/from-date/{date}:
    get:
      tags:
        - StockItemTemplate
      summary: Retrieves stock item templates from a specific date
      description: Fetches stock item templates recorded from a given date onward.
      parameters:
        - name: date
          in: path
          description: The date in YYYY-MM-DD format.
          required: true
          style: simple
          schema:
            type: string
            default: ""
            example: 2020-01-01
        - name: limit
          in: query
          description: "Maximum number of results to return (default: 100)."
          style: form
          schema:
            type: integer
            format: int32
            default: 100
            example: 1
        - name: offset
          in: query
          description: "Offset for pagination (default: 0)."
          style: form
          schema:
            type: integer
            format: int32
            default: 0
            example: 0
      responses:
        '200':
          description: Returns the details of the business object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockItemTemplate'
              examples:
                default_example:
                  $ref: '#/components/examples/StockItemTemplateInfoResponseExample'
        '400':
          description: General error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      additionalProperties: false
      description: Error response model.
      properties:
        ERRORMESSAGE:
          example: 'Login failed: Invalid user name or password'
          oneOf:
          - type: number
          - type: string
      type: object
    ItemTemplateDetail:
      additionalProperties: false
      properties:
        AMOUNT:
          example: 700
          oneOf:
          - type: number
          - type: string
        CODE_1:
          example: NewStockItem
          oneOf:
          - type: number
          - type: string
        DESCRIPTION_1:
          example: Premium PU Leather Ultimate Gaming Chair (Black)
          oneOf:
          - type: number
          - type: string
        DESCRIPTION2_1:
          example: ''
          oneOf:
          - type: number
          - type: string
        DESCRIPTION3_1:
          example: ''
          oneOf:
          - type: number
          - type: string
        DISC:
          example: ''
          oneOf:
          - type: number
          - type: string
        DTLKEY:
          example: 1
          oneOf:
          - type: number
          - type: string
        ITEMCODE:
          example: BOM-01/Black
          oneOf:
          - type: number
          - type: string
        LOCATION:
          example: '----'
          oneOf:
          - type: number
          - type: string
        NUMBER:
          example: ''
          oneOf:
          - type: number
          - type: string
        PRINTABLE:
          example: T
          oneOf:
          - type: number
          - type: string
        PROJECT:
          example: '----'
          oneOf:
          - type: number
          - type: string
        QTY:
          example: 1
          oneOf:
          - type: number
          - type: string
        REMARK_1:
          example: ''
          oneOf:
          - type: number
          - type: string
        REMARK2:
          example: ''
          oneOf:
          - type: number
          - type: string
        SEQ:
          example: 1000
          oneOf:
          - type: number
          - type: string
        STYLEID:
          example: ''
          oneOf:
          - type: number
          - type: string
        SUOMQTY:
          example: 0
          oneOf:
          - type: number
          - type: string
        UNITAMOUNT:
          example: 700
          oneOf:
          - type: number
          - type: string
        UOM:
          example: UNIT
          oneOf:
          - type: number
          - type: string
      type: object
...
    StockItem:
      additionalProperties: false
      properties:
        ATTACHMENTS:
          example: ''
          oneOf:
          - type: number
          - type: string
        BALSQTY:
          example: -5
          oneOf:
          - type: number
          - type: string
        BALSUOMQTY:
          example: 0
          oneOf:
          - type: number
          - type: string
        BOM_ASMCOST:
          example: 0
          oneOf:
          - type: number
          - type: string
        BOM_LEADTIME:
          example: 0
          oneOf:
          - type: number
          - type: string
        cdsUOM:
          items:
            $ref: '#/components/schemas/UOM'
          type: array
        CODE:
          example: SEMI BOM
          oneOf:
          - type: number
          - type: string
        COSTINGMETHOD:
          example: 1
          oneOf:
          - type: number
          - type: string
        CREATIONDATE:
          example: 2024-08-08T00
          oneOf:
          - type: number
          - type: string
        DEFUOM_PH:
          example: ''
          oneOf:
          - type: number
          - type: string
        DEFUOM_SL:
          example: ''
          oneOf:
          - type: number
          - type: string
        DEFUOM_ST:
          example: ''
          oneOf:
          - type: number
          - type: string
        DESCRIPTION:
          example: SEMI BOM
          oneOf:
          - type: number
          - type: string
        DESCRIPTION2:
          example: ''
          oneOf:
          - type: number
          - type: string
        DESCRIPTION3:
          example: ''
          oneOf:
          - type: number
          - type: string
        DOCKEY:
          example: 1
          oneOf:
          - type: number
          - type: string
        IRBM_CLASSIFICATION:
          example: 18
          oneOf:
          - type: number
          - type: string
        ISACTIVE:
          example: T
          oneOf:
          - type: number
          - type: string
        ITEMTYPE:
          example: B
          oneOf:
          - type: number
          - type: string
        LASTMODIFIED:
          example: 1723112180
          format: int64
          type: integer
        LEADTIME:
          example: 0
          oneOf:
          - type: number
          - type: string
        MAXQTY:
          example: 0
          oneOf:
          - type: number
          - type: string
        MINQTY:
          example: 0
          oneOf:
          - type: number
          - type: string
        NOTE:
          example: ''
          oneOf:
          - type: number
          - type: string
        PHTAX:
          example: ''
          oneOf:
          - type: number
          - type: string
        PICTURE:
          example: ''
          oneOf:
          - type: number
          - type: string
        PICTURECLASS:
          example: ''
          oneOf:
          - type: number
          - type: string
        REMARK_1:
          example: ''
          oneOf:
          - type: number
          - type: string
        REMARK2:
          example: ''
          oneOf:
          - type: number
          - type: string
        REORDERLEVEL:
          example: 0
          oneOf:
          - type: number
          - type: string
        REORDERQTY:
          example: 1
          oneOf:
          - type: number
          - type: string
        SCRIPTCODE:
          example: ''
          oneOf:
          - type: number
          - type: string
        SERIALNUMBER:
          example: F
          oneOf:
          - type: number
          - type: string
        SHELF:
          example: ''
          oneOf:
          - type: number
          - type: string
        SLTAX:
          example: ''
          oneOf:
          - type: number
          - type: string
        STOCKCONTROL:
          example: T
          oneOf:
          - type: number
          - type: string
        STOCKGROUP:
          example: OTHERS
          oneOf:
          - type: number
          - type: string
        STOCKMATRIX:
          example: ''
          oneOf:
          - type: number
          - type: string
        SUOM:
          example: ''
          oneOf:
          - type: number
          - type: string
        TARIFF:
          example: ''
          oneOf:
          - type: number
          - type: string
      type: object
...
    StockItemTemplate:
      additionalProperties: false
      properties:
        ATTACHMENTS:
          example: ''
          oneOf:
          - type: number
          - type: string
        cdsItemTplDtl:
          items:
            $ref: '#/components/schemas/ItemTemplateDetail'
          type: array
        CODE:
          example: NewStockItem
          oneOf:
          - type: number
          - type: string
        DESCRIPTION:
          example: ''
          oneOf:
          - type: number
          - type: string
        DESCRIPTION2:
          example: ''
          oneOf:
          - type: number
          - type: string
        DESCRIPTION3:
          example: ''
          oneOf:
          - type: number
          - type: string
        ISACTIVE:
          example: T
          oneOf:
          - type: number
          - type: string
        LASTMODIFIED:
          example: 1741233335
          format: int64
          type: integer
        REFPRICE:
          example: 0
          oneOf:
          - type: number
          - type: string
      type: object
...
    UOM:
      additionalProperties: false
      properties:
        CODE_1:
          example: SEMI BOM
          oneOf:
          - type: number
          - type: string
        ISBASE:
          example: 1
          oneOf:
          - type: number
          - type: string
        MAXCOST:
          example: ''
          oneOf:
          - type: number
          - type: string
        MAXPRICE:
          example: ''
          oneOf:
          - type: number
          - type: string
        MINCOST:
          example: ''
          oneOf:
          - type: number
          - type: string
        MINPRICE:
          example: ''
          oneOf:
          - type: number
          - type: string
        RATE:
          example: 1
          oneOf:
          - type: number
          - type: string
        REFCOST:
          example: 23
          oneOf:
          - type: number
          - type: string
        REFPRICE:
          example: 0
          oneOf:
          - type: number
          - type: string
        UOM:
          example: UNIT
          oneOf:
          - type: number
          - type: string
      type: object
  examples:
...
    StockItemInfoResponseExample:
      value:
      - DOCKEY: 1
        CODE: SEMI BOM
        DESCRIPTION: SEMI BOM
        DESCRIPTION2: ""
        DESCRIPTION3: ""
        STOCKGROUP: OTHERS
        STOCKCONTROL: T
        COSTINGMETHOD: 1
        SERIALNUMBER: F
        REMARK_1: ""
        REMARK2: ""
        MINQTY: 0
        MAXQTY: 0
        REORDERLEVEL: 0
        REORDERQTY: 1
        SHELF: ""
        SUOM: ""
        ITEMTYPE: B
        LEADTIME: 0
        BOM_LEADTIME: 0
        BOM_ASMCOST: 0
        SLTAX: ""
        PHTAX: ""
        TARIFF: ""
        IRBM_CLASSIFICATION: 022
        STOCKMATRIX: ""
        DEFUOM_ST: ""
        DEFUOM_SL: ""
        DEFUOM_PH: ""
        SCRIPTCODE: ""
        ISACTIVE: T
        BALSQTY: -5
        BALSUOMQTY: 0
        CREATIONDATE: 2024-08-08T00:00:00
        PICTURE: ""
        PICTURECLASS: ""
        ATTACHMENTS: ""
        NOTE: ""
        LASTMODIFIED: 1723112180
        cdsUOM:
        - CODE_1: SEMI BOM
          UOM: UNIT
          RATE: 1
          REFCOST: 23
          REFPRICE: 0
          MINCOST: ""
          MAXCOST: ""
          MINPRICE: ""
          MAXPRICE: ""
          ISBASE: 1
    StockItemTemplateInfoResponseExample:
      value:
      - CODE: NewStockItem
        DESCRIPTION: ""
        DESCRIPTION2: ""
        DESCRIPTION3: ""
        REFPRICE: ""
        ISACTIVE: T
        ATTACHMENTS: ""
        LASTMODIFIED: 1741233335
        cdsItemTplDtl:
        - DTLKEY: 1
          CODE_1: NewStockItem
          SEQ: 1000
          STYLEID: ""
          NUMBER: ""
          ITEMCODE: BOM-01/Black
          LOCATION: ----
          PROJECT: ----
          DESCRIPTION_1: Premium PU Leather Ultimate Gaming Chair (Black)
          DESCRIPTION2_1: ""
          DESCRIPTION3_1: ""
          QTY: 1
          SUOMQTY: 0
          UOM: UNIT
          UNITAMOUNT: 700
          DISC: ""
          AMOUNT: 700
          PRINTABLE: T
          REMARK_1: ""
          REMARK2: ""
...